get-data/requirements.txt
google-cloud-bigquery==3.4.0
python-dotenv==0.19.0
google-auth==2.14.1
db-dtypes==1.0.4
google-cloud-storage==2.6.0